fix bug in gemv

+++ GEMV/Sgemv_v2.cu
@@ -5,7 +5,6 @@
 #include <cuda_runtime.h>
 #include <cublas_v2.h>
 #include <math.h> 
-#include "cuHalfComplex.h"
 
 // cal offset from row col and ld , in row-major matrix, ld is the width of the matrix
 #define OFFSET(row, col, ld) ((row) * (ld) + (col))
@@ -95,7 +94,7 @@ int main(int argc, char** argv) {
 
     // 生成A的数据
     for( int i = 0; i < M * N; i++ ) {
-        h_A[i] = i/N;
+        h_A[i] = (float)i/N;
     }
 
     // 生成x的数据
